video-tool/requirements.txt
streamlit>=1.30.0
PyVimeo>=1.0.10
python-dotenv>=1.0.0
Pillow>=10.0.0
